+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 02/10/2023 - v4.0.2
+
+- (Change) UI: Layer height inputs minimum from 0.01mm to 0.001mm (#761)
+- (Fix) Goo: The value of "RetractHeight" was not being set to the file (#763)
+
 ## 18/09/2023 - v4.0.1
 
 - (Fix) Suggestions settings were not displaying