+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
-- (Fix) Suggestions settings was not displaying
+- (Change) UI: Layer height inputs minimum from 0.01mm to 0.001mm (#761)
+- (Fix) Goo: The value of "RetractHeight" was not being set to the file (#763)